+++ Questionnaire/Questionnaire.BLL/DTO/QuestionDTO.cs
@@ -5,7 +5,7 @@ namespace Questionnaire.BLL.DTO
     public class QuestionDTO
     {
         public int Id { get; set; }
-        public int QuestionTypeId { get; set; }
+        public short QuestionTypeId { get; set; }
         public string Name { get; set; }
 
         public QuestionTypeDTO QuestionType { get; set; }
