+++ Questionnaire/Questionnaire.DAL/Entities/Form.cs
@@ -7,10 +7,10 @@ namespace Questionnaire.DAL.Entities
     public class Form
     {
         public int Id { get; set; }
-        public int SurveyGeographyId { get; set; }
-        public int HousingTypeId { get; set; }
-        public int DistrictId { get; set; }    
-        public int InterviewerId { get; set; }       
+        public short SurveyGeographyId { get; set; }
+        public short HousingTypeId { get; set; }
+        public short DistrictId { get; set; }    
+        public short InterviewerId { get; set; }       
         public string Address { get; set; }
         public string Phone { get; set; }
         public DateTime InterviewDate { get; set; }
